mapper, support: restructure block module
